iteration 2
In this iteration 2, I tried to enrich the canvas with images of stars and the moon. The new code includes the loadImage function to load the star and moon images. At the same time, I adjusted the size and position of these images so that their position in the sky fits the width of the canvas.