iteration 3 (final)
In the final iteration, I made many changes because I found a bug in the code of iteration 2. When the oil painting image is loaded, mousePressed() cannot be used. Therefore, I chose to let the oil painting automatically jump to the original image, so I removed noLoop(). That the draw() function can continue to execute when mousePressed() is called. I also added clear() to automatically jump to the original image after loading the oil painting image, and then added the filter and star image after mousePressed().

In addition, I added animated stars to increase the visual look. I referred to the Star example of https://editor.p5js.org/p5/sketches/Form:_Star. The painting state is saved and restored through push() and pop(), and the stars are rotated using translate() and rotate().